#687: adds typedocs for shellbar component
