--- library/src/lib/shellbar/copilot.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'fd-copilot',
-    templateUrl: './copilot.component.html'
-})
-export class CopilotComponent {}
