+++ library/src/lib/shellbar/copilot/copilot.component.html
