+++ library/src/lib/shellbar/copilot/copilot.component.spec.ts
