+++ library/src/lib/shellbar/copilot/copilot.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+
+/**
+ * The component that represents a CoPilot.
+ * ```html
+ * <fd-copilot></fd-copilot>
+ * ```
+ */
+@Component({
+    selector: 'fd-copilot',
+    templateUrl: './copilot.component.html'
+})
+export class CopilotComponent { }
