+++ library/src/lib/shellbar/product-menu/product-menu.component.html
