+++ library/src/lib/shellbar/product-menu/product-menu.component.spec.ts
@@ -1,8 +1,8 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ProductMenuComponent } from './product-menu.component';
-import { PopoverModule } from '../popover/popover.module';
-import { MenuModule } from '../menu/menu.module';
+import { PopoverModule } from '../../popover/popover.module';
+import { MenuModule } from '../../menu/menu.module';
 
 describe('ProductMenuComponent', () => {
     let component: ProductMenuComponent;
