+++ library/src/lib/shellbar/shellbar-action/shellbar-action.component.ts
@@ -1,0 +1,46 @@
+import { Component, ContentChild, Input } from '@angular/core';
+import { SearchInputComponent } from '../../search-input/search-input.component';
+
+/**
+ * The component that represents a shellbar action.
+ * ```html
+ *  <fd-shellbar-action *ngFor="let action of actions"
+ *                      [glyph]="action.glyph"
+ *                      [callback]="action.callback"
+ *                      [label]="action.label"
+ *                      [notificationCount]="action.notificationCount"
+ *                      [notificationLabel]="action.notificationLabel">
+ *  </fd-shellbar-action>
+ * ```
+ */
+@Component({
+    selector: 'fd-shellbar-action',
+    templateUrl: './shellbar-action.component.html'
+})
+export class ShellbarActionComponent {
+
+    /** The glyph (icon) name */
+    @Input()
+    glyph: string;
+
+    /** Callback that hanldles the response to clicks on any of the actions. */
+    @Input()
+    callback: Function;
+
+    /** The action label. */
+    @Input()
+    label: string;
+
+    /** The notification label. */
+    @Input()
+    notificationLabel: string;
+
+    /** Represents the number of notifications. */
+    @Input()
+    notificationCount: number;
+
+    /** @hidden */
+    @ContentChild(SearchInputComponent)
+    searchInputComponent: SearchInputComponent;
+
+}
