--- library/src/lib/shellbar/shellbar-logo.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'fd-shellbar-logo',
-    templateUrl: './shellbar-logo.component.html'
-})
-export class ShellbarLogoComponent {}
