+++ library/src/lib/shellbar/shellbar-logo/shellbar-logo.component.html
