--- library/src/lib/shellbar/shellbar-subtitle.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'fd-shellbar-subtitle',
-    templateUrl: './shellbar-subtitle.component.html'
-})
-export class ShellbarSubtitleComponent {}
