--- library/src/lib/shellbar/shellbar-title.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'fd-shellbar-title',
-    templateUrl: './shellbar-title.component.html'
-})
-export class ShellbarTitleComponent {}
