+++ library/src/lib/shellbar/shellbar-title/shellbar-title.component.ts
@@ -1,0 +1,16 @@
+import { Component } from '@angular/core';
+
+/**
+ * The component that represents a shellbar title.
+ * The title is a required element and displays the current application name.
+ * ```html
+ *   <fd-shellbar-title>
+ *      Corporate Portal
+ *   </fd-shellbar-title>
+ * ```
+ */
+@Component({
+    selector: 'fd-shellbar-title',
+    templateUrl: './shellbar-title.component.html'
+})
+export class ShellbarTitleComponent { }
