+++ library/src/lib/shellbar/shellbar.component.ts
@@ -1,5 +1,10 @@
 import { Component } from '@angular/core';
 
+/**
+ * The shellbar offers consistent, responsive navigation across all products and applications. 
+ * Includes support for branding, product navigation, search, notifications, user settings, and CoPilot. 
+ * Shellbar is a composite component comprised of mandatory and optional elements. 
+ */
 @Component({
   selector: 'fd-shellbar',
   templateUrl: './shellbar.component.html'
